MongoDB implemented with maven compilation
