--- MongoDB/build.gradle
@@ -1,32 +1,0 @@
-apply plugin: 'java'
-apply plugin: 'application'
-
-startScripts.enabled = false
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-  compile 'org.mongodb:mongodb-driver:3.2.2'
-}
-
-task MDBJavaSignup(type: CreateStartScripts) {
-  mainClassName = "Maalgm.MDBJavaSignup"
-  applicationName = "mdb-java-signup"
-  outputDir = new File(project.buildDir, 'tmp')
-  classpath = jar.outputs.files + project.configurations.runtime
-}
-
-task MDBJavaLogin(type: CreateStartScripts) {
-  mainClassName = "Maalgm.MDBJavaLogin"
-  applicationName = "mdb-java-login"
-  outputDir = new File(project.buildDir, 'tmp')
-  classpath = jar.outputs.files + project.configurations.runtime
-}
-
-applicationDistribution.into("bin") {
-  from(MDBJavaSignup)
-  from(MDBJavaLogin)
-  fileMode = 0755
-}
