+++ README.md
@@ -1,1 +1,6 @@
-# Maalgm
+# Maalgm MongoDB implementation
+
+How to build & install: inside directory 'MDBJava', use 'mvn package' and 'mvn install -Dmaven.test.skip=true'
+How to execute: 'mvn exec:java -Dexec.mainClass="maalgm.(java class name)" -Dexec.classpathScope=runtime'
+
+Test files are planned to added after connection with other implementations.
